Add files via upload
Model for the inclusion of extra features to be used in a supervised learning algorithm of our choice. I've come up with features. Any other ideas are welcome. If we can get the cosine similarity of all the questions given the new query (question), then we can add all of these other features in. This also gives us something to work with as far as tasks for individuals to build out.